+++ CHANGES.md
@@ -9,6 +9,7 @@
 - **Generell**: Den Komponenten wird zur übersicht und erweiterbarkeit der gesamte Controller übergeben
 - **Generell**: Hinzufügen eines SettingsButtonComponent welcher zwischen DroneEditor und Settings switcht
 
+- **DroneEditorComponent**: colorMap entfernt
 - **SettingsComponent**: _droneDistance_ geändert zu _collisionRadius_
 - **EditorComponent**: simulation-view szene im Interface hinzugefügt